moved geometry config and heap commitments
