+++ circuit_sequencer_api/Cargo.toml
@@ -9,7 +9,7 @@ license = "MIT OR Apache-2.0"
 [dependencies]
 
 zk_evm = {git = "https://github.com/matter-labs/era-zk_evm.git", branch = "v1.4.0"}
-circuit_definitions = {path = "../circuit_definitions"}
+circuit_encodings = {path = "../circuit_encodings"}
 bellman = { package = "bellman_ce", git = "https://github.com/matter-labs/bellman", branch = "dev" }
