+++ circuit_sequencer_api/src/geometry_config.rs
