+++ circuit_sequencer_api/src/lib.rs
@@ -1,4 +1,13 @@
+#![feature(array_chunks)]
+
+pub mod geometry_config;
 pub mod proof;
 pub mod sort_storage_access;
+pub mod toolset;
+
+pub mod commitments;
+pub mod utils;
+
+pub use circuit_encodings::boojum;
 
 pub const INITIAL_MONOTONIC_CYCLE_COUNTER: u32 = 1024;
