+++ circuit_sequencer_api/src/sort_storage_access.rs
@@ -4,8 +4,8 @@ use std::cmp::Ordering;
 use std::iter::IntoIterator;
 use zk_evm::ethereum_types::U256;
 
-use circuit_definitions::encodings::LogQueryLike;
-use circuit_definitions::encodings::LogQueryLikeWithExtendedEnumeration;
+use circuit_encodings::LogQueryLike;
+use circuit_encodings::LogQueryLikeWithExtendedEnumeration;
 
 #[derive(Derivative)]
 #[derivative(Default(bound = ""), Debug)]
