+++ src/utils.rs
@@ -34,31 +34,6 @@ pub fn u128_as_u32_le(value: u128) -> [u32; 4] {
     ]
 }
 
-pub fn calldata_to_aligned_data(calldata: &Vec<u8>) -> Vec<U256> {
-    if calldata.len() == 0 {
-        return vec![];
-    }
-    let mut capacity = calldata.len() / 32;
-    if calldata.len() % 32 != 0 {
-        capacity += 1;
-    }
-    let mut result = Vec::with_capacity(capacity);
-    let mut it = calldata.chunks_exact(32);
-    for el in &mut it {
-        let el = U256::from_big_endian(el);
-        result.push(el);
-    }
-    let remainder = it.remainder();
-    if remainder.len() != 0 {
-        let mut buffer = [0u8; 32];
-        buffer[0..remainder.len()].copy_from_slice(remainder);
-        let el = U256::from_big_endian(&buffer);
-        result.push(el);
-    }
-
-    result
-}
-
 pub fn bytes_to_u32_le<const N: usize, const M: usize>(bytes: &[u8; N]) -> [u32; M] {
     assert!(M > 0);
     assert!(M * 4 == N);
@@ -132,45 +107,6 @@ pub fn binary_merklize_set<
     root
 }
 
-use crate::boojum::algebraic_props::round_function::AlgebraicRoundFunction;
-use crate::boojum::gadgets::traits::round_function::BuildableCircuitRoundFunction;
-use crate::zkevm_circuits::scheduler::QUEUE_FINAL_STATE_COMMITMENT_LENGTH;
-use circuit_definitions::encodings::OutOfCircuitFixedLengthEncodable;
-
-pub fn finalize_queue_state<
-    F: SmallField,
-    R: BuildableCircuitRoundFunction<F, 8, 12, 4> + AlgebraicRoundFunction<F, 8, 12, 4>,
-    const N: usize,
->(
-    tail: [F; N],
-    _round_function: &R,
-) -> [F; QUEUE_FINAL_STATE_COMMITMENT_LENGTH] {
-    // rescue prime paddings
-    let mut to_absorb = vec![];
-    to_absorb.extend(tail);
-    to_absorb.push(F::ONE);
-
-    let mut state = R::initial_state();
-    use crate::boojum::algebraic_props::round_function::absorb_into_state_vararg;
-    absorb_into_state_vararg::<F, R, AbsorptionModeOverwrite, 8, 12, 4>(&mut state, &to_absorb);
-    let commitment = <R as AlgebraicRoundFunction<F, 8, 12, 4>>::state_into_commitment::<
-        QUEUE_FINAL_STATE_COMMITMENT_LENGTH,
-    >(&state);
-
-    commitment
-}
-
-pub fn finalized_queue_state_as_bytes<F: SmallField>(
-    input: [F; QUEUE_FINAL_STATE_COMMITMENT_LENGTH],
-) -> [u8; 32] {
-    let mut result = [0u8; 32];
-    for (dst, src) in result.array_chunks_mut::<8>().zip(input.into_iter()) {
-        *dst = src.as_u64_reduced().to_be_bytes();
-    }
-
-    result
-}
-
 use crate::boojum::pairing::bls12_381::fr::{Fr, FrRepr};
 use crate::sha2::{Digest, Sha256};
 use crate::sha3::Keccak256;
@@ -208,14 +144,12 @@ pub fn generate_eip4844_witness<F: SmallField>(
     // compute versioned hash
 
     // There chunks are representation of the monomial form
-    let mut poly = crate::zkevm_circuits::eip_4844::zksync_pubdata_into_monomial_form_poly(
-        &blob
-    );
+    let mut poly = crate::zkevm_circuits::eip_4844::zksync_pubdata_into_monomial_form_poly(&blob);
     // so FFT then
     crate::zkevm_circuits::eip_4844::fft(&mut poly);
     // and bitreverse
     crate::zkevm_circuits::eip_4844::bitreverse(&mut poly);
-    // now they can be an input to KZG commitment 
+    // now they can be an input to KZG commitment
 
     use crate::kzg::compute_commitment;
     use circuit_definitions::boojum::pairing::CurveAffine;
@@ -334,3 +268,7 @@ pub fn generate_eip4844_circuit_and_witness(
     };
     (circuit, witness)
 }
+
+pub use circuit_sequencer_api::utils::calldata_to_aligned_data;
+pub use circuit_sequencer_api::utils::finalize_queue_state;
+pub use circuit_sequencer_api::utils::finalized_queue_state_as_bytes;
